Configure webpack to report time of each compile...
With webpack 'hot-reload', there may be no feedback that webpack
noticed a file change and is serving the new file.
Searching for some way to configure webpack to report on each compile,
I found others discussing the issue in [webpack --watch needs timestamp]
(https://github.com/webpack/webpack/issues/1499).

Their solution is a simple plugin. So simple it can be code in the
config, as reported in [webpack-cli --watch needs timestamp]
(https://github.com/webpack/webpack-cli/issues/21) and included
in this commit.